+++ plasflow/exceptions.py
@@ -1,0 +1,8 @@
+
+
+class UnavailableLayerSpecError(Exception):
+    pass
+
+
+class UnavailableKmerSpecError(Exception):
+    pass
